chore: stronger api typing, mitigate "window is not defined"
When you run the app in dev mode, you get a warning about "window is not
defined" because `window` was accessed directly in the component body.

This commit adds better typing and an assertion function to avoid
checking the api version and using `in` every single time, similar to
the approach in the commmunity tool.
